requirements.txt
appdirs==1.4.3
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.1
idna==2.8
packaging==19.0
pyparsing==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
